AMP-7368
git-svn-id: file:///svn/trunk@10883 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/dbentity/AmpFundingDetail.java
@@ -38,7 +38,7 @@ public class AmpFundingDetail implements Serializable {
 	private String expCategory;
 	private String disbOrderId;
 	private IPAContract contract;
-	private Double transactionAmountInBaseCurrency;
+	private BigDecimal transactionAmountInBaseCurrency;
 
 	public AmpFundingDetail() {
 	}
@@ -78,7 +78,7 @@ public class AmpFundingDetail implements Serializable {
        /**
 		 * @return the transactionAmountInBaseCurrency
 		 */
-		public Double getTransactionAmountInBaseCurrency() {
+		public BigDecimal getTransactionAmountInBaseCurrency() {
 			return transactionAmountInBaseCurrency;
 		}
 
@@ -87,7 +87,7 @@ public class AmpFundingDetail implements Serializable {
 		 * @param transactionAmountInBaseCurrency the transactionAmountInBaseCurrency to set
 		 */
 		public void setTransactionAmountInBaseCurrency(
-				Double transactionAmountInBaseCurrency) {
+				BigDecimal transactionAmountInBaseCurrency) {
 			this.transactionAmountInBaseCurrency = transactionAmountInBaseCurrency;
 		}
